refactor(metrics): remove dead consumer & provider metrics managers
Only SmartRouterMetricsManager is instantiated at runtime. The consumer
(lava_consumer_*) and provider (lava_provider_*) metric families were never
registered: NewConsumerMetricsManager / NewProviderMetricsManager had no call
sites, and the ProviderMetricsManager type was plumbed through the chaintracker
only as an always-nil field.

Delete the dead code:
- consumer_metrics_manager.go, provider_metrics_manager.go, provider_metrics.go
  and provider_metrics_test.go
- the always-nil pmetrics field + its four nil-no-op SetXxxFetch calls in the
  chaintracker (the real fetch-error signal stays on fetchErrorCallback)
- the ConsumerMetricsManager-specific test cases across the shared metric test
  files (SmartRouterMetricsManager coverage is kept intact)

Relocate the load-bearing symbols that lived in the deleted files:
- flag-name constants + DisabledFlagOption + ShowProviderEndpointInMetrics
  -> new metrics/flags.go
- SetVersionInner (shared protocol-version encoder) -> new metrics/version_metric.go
- ProviderSelectionScores -> consumer_metrics_manager_inf.go (next to the live
  ConsumerMetricsManagerInf it feeds)

No runtime behaviour change: none of the removed metrics were ever exposed.
