+++ protocol/chaintracker/config.go
@@ -2,8 +2,6 @@ package chaintracker
 
 import (
 	"time"
-
-	"github.com/magma-Devs/smart-router/protocol/metrics"
 )
 
 const (
@@ -22,7 +20,6 @@ type ChainTrackerConfig struct {
 	AverageBlockTime         time.Duration // how often to query latest block
 	ServerBlockMemory        uint64
 	BlocksCheckpointDistance uint64 // this causes the chainTracker to trigger it's checkpoint every X blocks
-	Pmetrics                 *metrics.ProviderMetricsManager
 	PollingTimeMultiplier    int
 	ChainId                  string
 	ParseDirectiveEnabled    bool
