+++ protocol/metrics/request_group_metrics_test.go
@@ -9,26 +9,6 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-// newConsumerForRequestGroupTest constructs a minimal ConsumerMetricsManager
-// with fresh, unregistered counters. This avoids HTTP-mux and Prometheus
-// registry conflicts when running multiple test functions.
-func newConsumerForRequestGroupTest() *ConsumerMetricsManager {
-	reqLabels := []string{"spec", "apiInterface", "provider_address", "method"}
-	return &ConsumerMetricsManager{
-		// Fields used by SetRelayMetrics before the request-group block
-		totalCURequestedMetric: prometheus.NewCounterVec(prometheus.CounterOpts{Name: "t_cu"}, []string{"spec", "apiInterface"}),
-		// Request-group counters under test
-		requestsTotalMetric:      prometheus.NewCounterVec(prometheus.CounterOpts{Name: "t_req_total"}, reqLabels),
-		requestsSuccessMetric:    prometheus.NewCounterVec(prometheus.CounterOpts{Name: "t_req_success"}, reqLabels),
-		requestsFailedMetric:     prometheus.NewCounterVec(prometheus.CounterOpts{Name: "t_req_failed"}, reqLabels),
-		requestsReadMetric:       prometheus.NewCounterVec(prometheus.CounterOpts{Name: "t_req_read"}, reqLabels),
-		requestsWriteMetric:      prometheus.NewCounterVec(prometheus.CounterOpts{Name: "t_req_write"}, reqLabels),
-		requestsDebugTraceMetric: prometheus.NewCounterVec(prometheus.CounterOpts{Name: "t_req_debug"}, reqLabels),
-		requestsArchiveMetric:    prometheus.NewCounterVec(prometheus.CounterOpts{Name: "t_req_archive"}, reqLabels),
-		requestsBatchMetric:      prometheus.NewCounterVec(prometheus.CounterOpts{Name: "t_req_batch"}, reqLabels),
-	}
-}
-
 // newSmartRouterForRequestGroupTest constructs a minimal SmartRouterMetricsManager
 // with fresh, unregistered counters for testing RecordDirectRelayEnd logic.
 func newSmartRouterForRequestGroupTest() *SmartRouterMetricsManager {
@@ -55,7 +35,7 @@ func newSmartRouterForRequestGroupTest() *SmartRouterMetricsManager {
 }
 
 // requestGroupRunner is a thin adapter that lets the shared test case table
-// drive both ConsumerMetricsManager and SmartRouterMetricsManager.
+// drive SmartRouterMetricsManager.
 type requestGroupRunner struct {
 	// invoke dispatches a single relay through the manager under test.
 	// err==nil means success; err!=nil means failure.
@@ -69,25 +49,6 @@ type requestGroupRunner struct {
 	archive *prometheus.CounterVec
 	debug   *prometheus.CounterVec
 	batch   *prometheus.CounterVec
-}
-
-func newConsumerRequestGroupRunner() *requestGroupRunner {
-	cmm := newConsumerForRequestGroupTest()
-	return &requestGroupRunner{
-		invoke: func(r RelayMetrics, e error) {
-			r.ChainID, r.APIType, r.ProviderAddress = "ETH1", "jsonrpc", "provider1"
-			cmm.SetRelayMetrics(&r, e)
-		},
-		labels:  func(method string) []string { return []string{"ETH1", "jsonrpc", "provider1", method} },
-		total:   cmm.requestsTotalMetric,
-		success: cmm.requestsSuccessMetric,
-		failed:  cmm.requestsFailedMetric,
-		read:    cmm.requestsReadMetric,
-		write:   cmm.requestsWriteMetric,
-		archive: cmm.requestsArchiveMetric,
-		debug:   cmm.requestsDebugTraceMetric,
-		batch:   cmm.requestsBatchMetric,
-	}
 }
 
 func newSmartRouterRequestGroupRunner() *requestGroupRunner {
@@ -124,9 +85,8 @@ func (r *requestGroupRunner) assertCounters(t *testing.T, method string,
 	require.Equal(t, wantBatch, testutil.ToFloat64(r.batch.WithLabelValues(l...)), "batch")
 }
 
-// TestSetRelayMetrics covers the request-group counter logic for both
-// ConsumerMetricsManager (via SetRelayMetrics) and SmartRouterMetricsManager
-// (via RecordDirectRelayEnd) using the same table of cases.
+// TestSetRelayMetrics covers the request-group counter logic for
+// SmartRouterMetricsManager (via RecordDirectRelayEnd) using a table of cases.
 func TestSetRelayMetrics(t *testing.T) {
 	cases := []struct {
 		name        string
@@ -185,7 +145,6 @@ func TestSetRelayMetrics(t *testing.T) {
 		name string
 		new  func() *requestGroupRunner
 	}{
-		{"consumer", newConsumerRequestGroupRunner},
 		{"smart_router", newSmartRouterRequestGroupRunner},
 	}
 
@@ -205,48 +164,9 @@ func TestSetRelayMetrics(t *testing.T) {
 	}
 }
 
-func TestConsumerSetRelayMetrics_NilManager(t *testing.T) {
-	var cmm *ConsumerMetricsManager
-	require.NotPanics(t, func() {
-		cmm.SetRelayMetrics(&RelayMetrics{ChainID: "ETH1", Success: true}, nil)
-	})
-}
-
 func TestSmartRouterRecordDirectRelayEnd_NilManager(t *testing.T) {
 	var m *SmartRouterMetricsManager
 	require.NotPanics(t, func() {
 		m.RecordDirectRelayEnd("ETH1", "jsonrpc", "ep1", "method", 10, true, &RelayMetrics{})
 	})
-}
-
-// TestConsumerSetRelayMetrics_PartitionInvariant asserts the documented counting
-// invariants: batch+read+write==total and success+failed==total across a mixed
-// workload. debug_trace and archive are orthogonal sub-categories and are NOT
-// expected to sum to total.
-func TestConsumerSetRelayMetrics_PartitionInvariant(t *testing.T) {
-	cmm := newConsumerForRequestGroupTest()
-	l := []string{"ETH1", "jsonrpc", "p1", "m"}
-
-	send := func(rm RelayMetrics, err error) {
-		rm.ChainID, rm.APIType, rm.ProviderAddress, rm.ApiMethod = "ETH1", "jsonrpc", "p1", "m"
-		cmm.SetRelayMetrics(&rm, err)
-	}
-
-	send(RelayMetrics{IsWrite: false}, nil)                 // read, success
-	send(RelayMetrics{IsWrite: true}, nil)                  // write, success
-	send(RelayMetrics{IsBatch: true}, nil)                  // batch, success
-	send(RelayMetrics{IsArchive: true}, errors.New("e"))    // read+archive, failed
-	send(RelayMetrics{IsDebugTrace: true}, nil)             // read+debug, success
-	send(RelayMetrics{IsWrite: true, IsArchive: true}, nil) // write+archive, success
-
-	total := testutil.ToFloat64(cmm.requestsTotalMetric.WithLabelValues(l...))
-	read := testutil.ToFloat64(cmm.requestsReadMetric.WithLabelValues(l...))
-	write := testutil.ToFloat64(cmm.requestsWriteMetric.WithLabelValues(l...))
-	batch := testutil.ToFloat64(cmm.requestsBatchMetric.WithLabelValues(l...))
-	success := testutil.ToFloat64(cmm.requestsSuccessMetric.WithLabelValues(l...))
-	failed := testutil.ToFloat64(cmm.requestsFailedMetric.WithLabelValues(l...))
-
-	require.Equal(t, float64(6), total)
-	require.Equal(t, total, batch+read+write, "batch+read+write must equal total")
-	require.Equal(t, total, success+failed, "success+failed must equal total")
 }
